refactor: update imports to use util module for consistency
