+++ src/release/pushTags.ts
@@ -1,5 +1,5 @@
 import { execa } from "execa";
-import { getTagBranchName } from "../getTagVersion.js";
+import { getTagBranchName } from "../util/index.js";
 
 /**
  * Push the tags to origin
